[wikimapia] Fix loading more detailed tiles
Wikimapia tiles sometimes are mistakenly marked as having no tiles on
more detailed levels. If such tile was found in cache, it was displayed
for all more detailed zoom levels.
Now instead we check if tile coordinates are same as were requested and
if they differ, we mark that tile does not have more detailed tiles
and update it in cache.