panoramas - fix wikimedia photos window size.
In Chrome, when panoramas are above map, on small (mobile) screens, and
more than 9 photos available, then scroll is not enabled for buttons
and no right navigation arrow is visible.
The problem is that the middle layout column has minimum size of 470px.
I do not know where this value comes from, and it is not so in Firefox.